requirements.txt
fsspec
gcsfs
pyspark
google-cloud
google-cloud-bigquery
